update hdf classes
